refactor(app): services files get named for what they export
Four files were named for something other than their export, and one held
two independent maps where constant-maps asks for two files. The dungeons
ScenePlugins interface is deleted outright: it restated, in a file nothing
imported, the three members types/vue-phaserjs.d.ts already augments onto
the library's own interface.

The exclusion for two exports sharing closure state is stated as the rule
rather than the one file that first showed it — three more files match it,
and publishing their private state to satisfy one-export-per-file would
trade a file boundary for a wider one.

Closes app/services.
