+++ .agents/ledgers/file-organization.md
@@ -11,7 +11,7 @@ The question is where a thing lives and whether it exists twice — one export p
 | `app/shared/models/dungeons`                                               | 2026-09-06 |                                                                                              |
 | `app/shared/models` — `clicker`, `dashboard`, `dataset`, `flowchartEditor` | 2026-09-06 |                                                                                              |
 | `app/shared/models` — the rest                                             | 2026-09-06 | the small folders, several of them a single file                                             |
-| `app/services`                                                             | —          | models vs services vs utils vs constants; duplicate constants                                |
+| `app/services`                                                             | 2026-09-06 | a file is named for its export, and a second map is a second file                            |
 | `app/models/dungeons`                                                      | 2026-09-06 | a class hierarchy stays a model; a map and the behaviour it dispatches to do not             |
 | `app/models/resource`                                                      | 2026-09-06 | a command class stays a model; the map beside a form union does not                          |
 | `app/models/message`                                                       | 2026-09-06 | a type derived from a services map is a type-only import, not a move                         |
@@ -75,8 +75,9 @@ finding: one consumer does not earn a place in a shared package.
 
 - Generated barrels (`index.ts` from `ctix`) and `snapshot.json` — machine state.
 - Literals a postinstall-evaluated or JSON config must repeat, which the skill names as the one sanctioned duplication.
-- `getSynchronizedFunction.ts` exporting `waitForSynchronizedFunctions` beside it: the pair shares the pending
-  set through closure, so one-export-per-file cannot reach them without making that state a module global.
+- Two exports sharing module-private state through closure — a pending set, a cached promise, a code set, a
+  dispatch map. One-export-per-file cannot reach them without making that state a module global, which trades a
+  file boundary for a wider one.
 
 ## Next enforceable
