+++ packages/app/app/components/Anime/Azunyan.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { AzunyanExpressions } from "@/services/anime/azunyanExpressions";
+import { AzunyanExpressions } from "@/services/anime/AzunyanExpressions";
 </script>
 
 <template>
