+++ packages/app/app/services/dungeons/UI/bar/BarTextureMap.ts
@@ -14,11 +14,3 @@ export const BarTextureMap = {
     [BarOrigin.Right]: ImageKey.HealthBarRightCap,
   },
 } as const satisfies Record<BarType, Record<BarOrigin, ImageKey>>;
-
-// Drawn behind every bar whatever it measures, so this is not a `BarType` — a shadow is not a thing a bar can
-// Be, it is what sits under one
-export const BarShadowTextureMap = {
-  [BarOrigin.Left]: ImageKey.BarLeftCapShadow,
-  [BarOrigin.Middle]: ImageKey.BarMiddleShadow,
-  [BarOrigin.Right]: ImageKey.BarRightCapShadow,
-} as const satisfies Record<BarOrigin, ImageKey>;
