--- packages/app/app/services/dungeons/scene/ScenePlugins.ts
@@ -1,9 +1,0 @@
-import type { GridEngine } from "grid-engine";
-import type SliderPlugin from "phaser4-rex-plugins/plugins/slider-plugin";
-import type VirtualJoystickPlugin from "phaser4-rex-plugins/plugins/virtualjoystick-plugin";
-
-export interface ScenePlugins {
-  gridEngine: GridEngine;
-  sliderPlugin: SliderPlugin;
-  virtualJoystickPlugin: VirtualJoystickPlugin;
-}
